+++ src/Translations/Screen/SettingsScreen.ts
@@ -4,7 +4,6 @@ export type TranslationDTO = Record<LanguageTag, {
   'Language': string
   'Date and time': string
   'Themes': string
-  'Credentials': string
 }>
 
 export const R_Screen_Settings: TranslationDTO = {
@@ -12,12 +11,10 @@ export const R_Screen_Settings: TranslationDTO = {
     'Language': 'Language',
     'Date and time': 'Date and time',
     'Themes': 'Themes',
-    'Credentials': 'Credentials',
   },
   'pt-BR': {
     'Language': 'Idioma',
     'Date and time': 'Data e hora',
     'Themes': 'Temas',
-    'Credentials': 'Credenciais',
   },
 };