+++ src/Translations/scope/HomeScope.ts
@@ -4,6 +4,7 @@ type TranslationDTO = Record<LanguageTag, {
   'Home screen': string
 	'Want to exit?': string
   'Settings': string
+  'Credentials': string
 }>
 
 export const R_Scope_Home: TranslationDTO = {
@@ -11,10 +12,12 @@ export const R_Scope_Home: TranslationDTO = {
     'Home screen': 'Main Screen',
 		'Want to exit?': 'Want to exit?',
     'Settings': 'Settings',
+    'Credentials': 'Credentials',
   },
   'pt-BR': {
     'Home screen': 'Tela inicial',
 		'Want to exit?': 'Deseja sair?',
     'Settings': 'Configurações',
+    'Credentials': 'Credenciais',
   },
 };